fix: Request를 파싱하는 책임을 Parser에서 Request로 위임
